@@ -2,18 +2,16 @@ apply plugin: 'com.android.application'
 
 apply plugin: 'kotlin-android'
 
-apply plugin: 'kotlin-android-extensions'
-
 apply plugin: 'kotlin-kapt'
 
 android {
-    compileSdkVersion 30
+    compileSdkVersion 32
     defaultConfig {
         applicationId "com.parsiphal.loganshopdriverhelper2"
-        minSdkVersion 21
-        targetSdkVersion 30
-        versionCode 202112182
-        versionName "0.3.41"
+        minSdkVersion 22
+        targetSdkVersion 32
+        versionCode 202212061
+        versionName "0.3.42"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
@@ -27,6 +25,10 @@ android {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
+    buildFeatures {
+        viewBinding = true
+    }
+    namespace 'com.parsiphal.loganshopdriverhelper'
 }
 
 dependencies {
@@ -57,7 +59,7 @@ dependencies {
     // Moxy
     ////////////////////////////////////////////////////////////////////////////////////////////////
     implementation 'com.arello-mobile:moxy:1.5.6'
-    annotationProcessor 'com.arello-mobile:moxy-compiler:1.5.6'
+    kapt 'com.arello-mobile:moxy-compiler:1.5.6'
     kapt 'com.github.moxy-community:moxy-compiler:1.0.13'
     implementation 'com.github.moxy-community:moxy-androidx:1.0.13'
 }
