@@ -8,32 +8,35 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.AdapterView
 import android.widget.TextView
-
 import com.parsiphal.loganshopdriverhelper.R
 import com.parsiphal.loganshopdriverhelper.data.Cars
+import com.parsiphal.loganshopdriverhelper.databinding.FragmentShiftBinding
 import com.parsiphal.loganshopdriverhelper.prefs
-import kotlinx.android.synthetic.main.fragment_shift.*
 import moxy.MvpAppCompatFragment
 import java.lang.Exception
 import java.util.*
 
 class ShiftFragment : MvpAppCompatFragment() {
 
+    private var _binding: FragmentShiftBinding? = null
+    private val binding get() = _binding!!
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_shift, container, false)
+        _binding = FragmentShiftBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        shift_status_textView.text = when (prefs.status) {
+        binding.shiftStatusTextView.text = when (prefs.status) {
             0 -> resources.getString(R.string.driver)
             else -> resources.getString(R.string.newbie)
         }
         getData()
-        shift_car_spinner.onItemSelectedListener = (object : AdapterView.OnItemSelectedListener {
+        binding.shiftCarSpinner.onItemSelectedListener = (object : AdapterView.OnItemSelectedListener {
             override fun onNothingSelected(parent: AdapterView<*>?) {
 
             }
@@ -46,97 +49,97 @@ class ShiftFragment : MvpAppCompatFragment() {
             ) {
                 when (position) {
                     Cars.Largus.i -> {
-                        shift_fuel_morning_spinner_6.visibility = View.GONE
-                        shift_fuel_evening_spinner_6.visibility = View.GONE
-                        shift_fuel_morning_spinner_8.visibility = View.GONE
-                        shift_fuel_evening_spinner_8.visibility = View.GONE
-                        shift_fuel_morning_spinner_9.visibility = View.VISIBLE
-                        shift_fuel_evening_spinner_9.visibility = View.VISIBLE
+                        binding.shiftFuelMorningSpinner6.visibility = View.GONE
+                        binding.shiftFuelEveningSpinner6.visibility = View.GONE
+                        binding.shiftFuelMorningSpinner8.visibility = View.GONE
+                        binding.shiftFuelEveningSpinner8.visibility = View.GONE
+                        binding.shiftFuelMorningSpinner9.visibility = View.VISIBLE
+                        binding.shiftFuelEveningSpinner9.visibility = View.VISIBLE
                     }
                     Cars.Sandero.i -> {
-                        shift_fuel_morning_spinner_6.visibility = View.GONE
-                        shift_fuel_evening_spinner_6.visibility = View.GONE
-                        shift_fuel_morning_spinner_8.visibility = View.GONE
-                        shift_fuel_evening_spinner_8.visibility = View.GONE
-                        shift_fuel_morning_spinner_9.visibility = View.VISIBLE
-                        shift_fuel_evening_spinner_9.visibility = View.VISIBLE
+                        binding.shiftFuelMorningSpinner6.visibility = View.GONE
+                        binding.shiftFuelEveningSpinner6.visibility = View.GONE
+                        binding.shiftFuelMorningSpinner8.visibility = View.GONE
+                        binding.shiftFuelEveningSpinner8.visibility = View.GONE
+                        binding.shiftFuelMorningSpinner9.visibility = View.VISIBLE
+                        binding.shiftFuelEveningSpinner9.visibility = View.VISIBLE
                     }
                     Cars.XRay.i -> {
-                        shift_fuel_morning_spinner_6.visibility = View.GONE
-                        shift_fuel_evening_spinner_6.visibility = View.GONE
-                        shift_fuel_morning_spinner_8.visibility = View.VISIBLE
-                        shift_fuel_evening_spinner_8.visibility = View.VISIBLE
-                        shift_fuel_morning_spinner_9.visibility = View.GONE
-                        shift_fuel_evening_spinner_9.visibility = View.GONE
+                        binding.shiftFuelMorningSpinner6.visibility = View.GONE
+                        binding.shiftFuelEveningSpinner6.visibility = View.GONE
+                        binding.shiftFuelMorningSpinner8.visibility = View.VISIBLE
+                        binding.shiftFuelEveningSpinner8.visibility = View.VISIBLE
+                        binding.shiftFuelMorningSpinner9.visibility = View.GONE
+                        binding.shiftFuelEveningSpinner9.visibility = View.GONE
                     }
                     Cars.LargusNew.i -> {
-                        shift_fuel_morning_spinner_6.visibility = View.GONE
-                        shift_fuel_evening_spinner_6.visibility = View.GONE
-                        shift_fuel_morning_spinner_8.visibility = View.GONE
-                        shift_fuel_evening_spinner_8.visibility = View.GONE
-                        shift_fuel_morning_spinner_9.visibility = View.VISIBLE
-                        shift_fuel_evening_spinner_9.visibility = View.VISIBLE
+                        binding.shiftFuelMorningSpinner6.visibility = View.GONE
+                        binding.shiftFuelEveningSpinner6.visibility = View.GONE
+                        binding.shiftFuelMorningSpinner8.visibility = View.GONE
+                        binding.shiftFuelEveningSpinner8.visibility = View.GONE
+                        binding.shiftFuelMorningSpinner9.visibility = View.VISIBLE
+                        binding.shiftFuelEveningSpinner9.visibility = View.VISIBLE
                     }
                     Cars.VestaSW.i -> {
-                        shift_fuel_morning_spinner_6.visibility = View.VISIBLE
-                        shift_fuel_evening_spinner_6.visibility = View.VISIBLE
-                        shift_fuel_morning_spinner_8.visibility = View.GONE
-                        shift_fuel_evening_spinner_8.visibility = View.GONE
-                        shift_fuel_morning_spinner_9.visibility = View.GONE
-                        shift_fuel_evening_spinner_9.visibility = View.GONE
+                        binding.shiftFuelMorningSpinner6.visibility = View.VISIBLE
+                        binding.shiftFuelEveningSpinner6.visibility = View.VISIBLE
+                        binding.shiftFuelMorningSpinner8.visibility = View.GONE
+                        binding.shiftFuelEveningSpinner8.visibility = View.GONE
+                        binding.shiftFuelMorningSpinner9.visibility = View.GONE
+                        binding.shiftFuelEveningSpinner9.visibility = View.GONE
                     }
                 }
             }
         })
-        shift_write.setOnClickListener {
+        binding.shiftWrite.setOnClickListener {
             try {
-                prefs.family = shift_family_editText.text.toString()
-                prefs.bonus = shift_bonus_checkbox.isChecked
-                prefs.date = shift_date.text.toString()
-                prefs.region = shift_region_spinner.selectedItem.toString()
-                prefs.regionPosition = shift_region_spinner.selectedItemPosition
-                prefs.car = shift_car_spinner.selectedItem.toString()
-                prefs.carPosition = shift_car_spinner.selectedItemPosition
-                prefs.morningODO = shift_odo_morning.text.toString().toInt()
-                prefs.eveningODO = shift_odo_evening.text.toString().toInt()
-                when (shift_car_spinner.selectedItemPosition) {
+                prefs.family = binding.shiftFamilyEditText.text.toString()
+                prefs.bonus = binding.shiftBonusCheckbox.isChecked
+                prefs.date = binding.shiftDate.text.toString()
+                prefs.region = binding.shiftRegionSpinner.selectedItem.toString()
+                prefs.regionPosition = binding.shiftRegionSpinner.selectedItemPosition
+                prefs.car = binding.shiftCarSpinner.selectedItem.toString()
+                prefs.carPosition = binding.shiftCarSpinner.selectedItemPosition
+                prefs.morningODO = binding.shiftOdoMorning.text.toString().toInt()
+                prefs.eveningODO = binding.shiftOdoEvening.text.toString().toInt()
+                when (binding.shiftCarSpinner.selectedItemPosition) {
                     Cars.XRay.i -> {
-                        prefs.morningFuel = (shift_fuel_morning_spinner_8.selectedItemPosition + 1)
-                        prefs.eveningFuel = (shift_fuel_evening_spinner_8.selectedItemPosition + 1)
+                        prefs.morningFuel = (binding.shiftFuelMorningSpinner8.selectedItemPosition + 1)
+                        prefs.eveningFuel = (binding.shiftFuelEveningSpinner8.selectedItemPosition + 1)
                     }
                     Cars.VestaSW.i -> {
-                        prefs.morningFuel = (shift_fuel_morning_spinner_6.selectedItemPosition + 1)
-                        prefs.eveningFuel = (shift_fuel_evening_spinner_6.selectedItemPosition + 1)
+                        prefs.morningFuel = (binding.shiftFuelMorningSpinner6.selectedItemPosition + 1)
+                        prefs.eveningFuel = (binding.shiftFuelEveningSpinner6.selectedItemPosition + 1)
                     }
                     else -> {
-                        prefs.morningFuel = (shift_fuel_morning_spinner_9.selectedItemPosition + 1)
-                        prefs.eveningFuel = (shift_fuel_evening_spinner_9.selectedItemPosition + 1)
+                        prefs.morningFuel = (binding.shiftFuelMorningSpinner9.selectedItemPosition + 1)
+                        prefs.eveningFuel = (binding.shiftFuelEveningSpinner9.selectedItemPosition + 1)
                     }
                 }
-                when (shift_car_spinner.selectedItemPosition) {
+                when (binding.shiftCarSpinner.selectedItemPosition) {
                     Cars.Largus.i -> {
                         prefs.lastFuelLargus =
-                            (shift_fuel_evening_spinner_9.selectedItemPosition + 1)
-                        prefs.lastOdoLargus = shift_odo_evening.text.toString().toInt()
+                            (binding.shiftFuelEveningSpinner9.selectedItemPosition + 1)
+                        prefs.lastOdoLargus = binding.shiftOdoEvening.text.toString().toInt()
                     }
                     Cars.Sandero.i -> {
                         prefs.lastFuelSandero =
-                            (shift_fuel_evening_spinner_9.selectedItemPosition + 1)
-                        prefs.lastOdoSandero = shift_odo_evening.text.toString().toInt()
+                            (binding.shiftFuelEveningSpinner9.selectedItemPosition + 1)
+                        prefs.lastOdoSandero = binding.shiftOdoEvening.text.toString().toInt()
                     }
                     Cars.XRay.i -> {
-                        prefs.lastFuelXRay = (shift_fuel_evening_spinner_8.selectedItemPosition + 1)
-                        prefs.lastOdoXRay = shift_odo_evening.text.toString().toInt()
+                        prefs.lastFuelXRay = (binding.shiftFuelEveningSpinner8.selectedItemPosition + 1)
+                        prefs.lastOdoXRay = binding.shiftOdoEvening.text.toString().toInt()
                     }
                     Cars.LargusNew.i -> {
                         prefs.lastFuelLargusNew =
-                            (shift_fuel_evening_spinner_9.selectedItemPosition + 1)
-                        prefs.lastOdoLargusNew = shift_odo_evening.text.toString().toInt()
+                            (binding.shiftFuelEveningSpinner9.selectedItemPosition + 1)
+                        prefs.lastOdoLargusNew = binding.shiftOdoEvening.text.toString().toInt()
                     }
                     Cars.VestaSW.i -> {
                         prefs.lastFuelVestaSW =
-                            (shift_fuel_evening_spinner_9.selectedItemPosition + 1)
-                        prefs.lastOdoVestaSW = shift_odo_evening.text.toString().toInt()
+                            (binding.shiftFuelEveningSpinner9.selectedItemPosition + 1)
+                        prefs.lastOdoVestaSW = binding.shiftOdoEvening.text.toString().toInt()
                     }
                 }
                 Snackbar.make(view, getString(R.string.recorded), Snackbar.LENGTH_SHORT).show()
@@ -145,36 +148,36 @@ class ShiftFragment : MvpAppCompatFragment() {
                 Snackbar.make(view, getString(R.string.wrongData), Snackbar.LENGTH_SHORT).show()
             }
         }
-        shift_date.setOnClickListener {
+        binding.shiftDate.setOnClickListener {
             datePickerDialog(it)
         }
-        shift_new_button.setOnClickListener {
+        binding.shiftNewButton.setOnClickListener {
             try {
-                when (shift_car_spinner.selectedItemPosition) {
+                when (binding.shiftCarSpinner.selectedItemPosition) {
                     Cars.Largus.i -> {
-                        shift_odo_morning.setText(prefs.lastOdoLargus!!.toString())
-                        shift_odo_evening.setText("0")
-                        shift_fuel_morning_spinner_9.setSelection(prefs.lastFuelLargus!!.minus(1))
+                        binding.shiftOdoMorning.setText(prefs.lastOdoLargus!!.toString())
+                        binding.shiftOdoEvening.setText("0")
+                        binding.shiftFuelMorningSpinner9.setSelection(prefs.lastFuelLargus!!.minus(1))
                     }
                     Cars.Sandero.i -> {
-                        shift_odo_morning.setText(prefs.lastOdoSandero!!.toString())
-                        shift_odo_evening.setText("0")
-                        shift_fuel_morning_spinner_9.setSelection(prefs.lastFuelSandero!!.minus(1))
+                        binding.shiftOdoMorning.setText(prefs.lastOdoSandero!!.toString())
+                        binding.shiftOdoEvening.setText("0")
+                        binding.shiftFuelMorningSpinner9.setSelection(prefs.lastFuelSandero!!.minus(1))
                     }
                     Cars.XRay.i -> {
-                        shift_odo_morning.setText(prefs.lastOdoXRay!!.toString())
-                        shift_odo_evening.setText("0")
-                        shift_fuel_morning_spinner_8.setSelection(prefs.lastFuelXRay!!.minus(1))
+                        binding.shiftOdoMorning.setText(prefs.lastOdoXRay!!.toString())
+                        binding.shiftOdoEvening.setText("0")
+                        binding.shiftFuelMorningSpinner8.setSelection(prefs.lastFuelXRay!!.minus(1))
                     }
                     Cars.LargusNew.i -> {
-                        shift_odo_morning.setText(prefs.lastOdoLargusNew!!.toString())
-                        shift_odo_evening.setText("0")
-                        shift_fuel_morning_spinner_9.setSelection(prefs.lastFuelLargusNew!!.minus(1))
+                        binding.shiftOdoMorning.setText(prefs.lastOdoLargusNew!!.toString())
+                        binding.shiftOdoEvening.setText("0")
+                        binding.shiftFuelMorningSpinner9.setSelection(prefs.lastFuelLargusNew!!.minus(1))
                     }
                     Cars.VestaSW.i -> {
-                        shift_odo_morning.setText(prefs.lastOdoVestaSW!!.toString())
-                        shift_odo_evening.setText("0")
-                        shift_fuel_morning_spinner_9.setSelection(prefs.lastFuelVestaSW!!.minus(1))
+                        binding.shiftOdoMorning.setText(prefs.lastOdoVestaSW!!.toString())
+                        binding.shiftOdoEvening.setText("0")
+                        binding.shiftFuelMorningSpinner9.setSelection(prefs.lastFuelVestaSW!!.minus(1))
                     }
                 }
 
@@ -186,41 +189,46 @@ class ShiftFragment : MvpAppCompatFragment() {
         }
     }
 
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
     private fun getData() {
         try {
-            shift_family_editText.setText(prefs.family)
-            shift_bonus_checkbox.isChecked = prefs.bonus
-            shift_date.text = prefs.date
-            shift_region_spinner.setSelection(prefs.regionPosition!!)
-            shift_car_spinner.setSelection(prefs.carPosition!!)
-            shift_odo_morning.setText(prefs.morningODO!!.toString())
-            shift_odo_evening.setText(prefs.eveningODO!!.toString())
-            when (shift_car_spinner.selectedItemPosition) {
+            binding.shiftFamilyEditText.setText(prefs.family)
+            binding.shiftBonusCheckbox.isChecked = prefs.bonus
+            binding.shiftDate.text = prefs.date
+            binding.shiftRegionSpinner.setSelection(prefs.regionPosition!!)
+            binding.shiftCarSpinner.setSelection(prefs.carPosition!!)
+            binding.shiftOdoMorning.setText(prefs.morningODO!!.toString())
+            binding.shiftOdoEvening.setText(prefs.eveningODO!!.toString())
+            when (binding.shiftCarSpinner.selectedItemPosition) {
                 Cars.XRay.i -> {
-                    shift_fuel_morning_spinner_8.setSelection(prefs.morningFuel!!.minus(1))
-                    shift_fuel_evening_spinner_8.setSelection(prefs.eveningFuel!!.minus(1))
+                    binding.shiftFuelMorningSpinner8.setSelection(prefs.morningFuel!!.minus(1))
+                    binding.shiftFuelEveningSpinner8.setSelection(prefs.eveningFuel!!.minus(1))
                 }
                 Cars.VestaSW.i -> {
-                    shift_fuel_morning_spinner_6.setSelection(prefs.morningFuel!!.minus(1))
-                    shift_fuel_evening_spinner_6.setSelection(prefs.eveningFuel!!.minus(1))
+                    binding.shiftFuelMorningSpinner6.setSelection(prefs.morningFuel!!.minus(1))
+                    binding.shiftFuelEveningSpinner6.setSelection(prefs.eveningFuel!!.minus(1))
                 }
                 else -> {
-                    shift_fuel_morning_spinner_9.setSelection(prefs.morningFuel!!.minus(1))
-                    shift_fuel_evening_spinner_9.setSelection(prefs.eveningFuel!!.minus(1))
+                    binding.shiftFuelMorningSpinner9.setSelection(prefs.morningFuel!!.minus(1))
+                    binding.shiftFuelEveningSpinner9.setSelection(prefs.eveningFuel!!.minus(1))
                 }
             }
         } catch (e: Exception) {
-            shift_family_editText.setText("")
-            shift_bonus_checkbox.isChecked = false
-            shift_date.text = ""
-            shift_region_spinner.setSelection(0)
-            shift_car_spinner.setSelection(0)
-            shift_odo_morning.setText("")
-            shift_odo_evening.setText("")
-            shift_fuel_morning_spinner_8.setSelection(0)
-            shift_fuel_evening_spinner_8.setSelection(0)
-            shift_fuel_morning_spinner_9.setSelection(0)
-            shift_fuel_evening_spinner_9.setSelection(0)
+            binding.shiftFamilyEditText.setText("")
+            binding.shiftBonusCheckbox.isChecked = false
+            binding.shiftDate.text = ""
+            binding.shiftRegionSpinner.setSelection(0)
+            binding.shiftCarSpinner.setSelection(0)
+            binding.shiftOdoMorning.setText("")
+            binding.shiftOdoEvening.setText("")
+            binding.shiftFuelMorningSpinner8.setSelection(0)
+            binding.shiftFuelEveningSpinner8.setSelection(0)
+            binding.shiftFuelMorningSpinner9.setSelection(0)
+            binding.shiftFuelEveningSpinner9.setSelection(0)
         }
     }
 
